Split tests and helper functions into files
